move(home): homepage screen 레이어로 이동
